+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2026.05.10
+
+### Added
+
+- Dumps AST to file or stderr
+- Compiler option for dump AST (--dump-ast=<none/file/term>)
+- Converting types to string (virtual std::string Type::ToString() method)
+
 ## [0.5.0] - 2026.05.09
 
 ### Added